% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beg_distribution.R
\name{bexpgeo_fit}
\alias{bexpgeo_fit}
\title{Fits the bivariate exponential geometric distribution (BEG) to data}
\usage{
bexpgeo_fit(data, level = 0.95)
}
\arguments{
\item{data}{is bivariate vector  (X,N) vector representing observations from BEG model}

\item{level}{confidence level expressed between 0 and 1 (Default is 0.95)}
}
\value{
list of parameter estimates, confidence interval, deviance and covariance matrix
}
\description{
This function computes the parameter estimates, confidence interval, deviance and covariance matrix for exponential geometric distribution.
}
\details{
bexpgeo_fit  fits BEG model to data.
}
\examples{
Data.df<-rbexpgeo(n=100,beta = 10,p=0.45)
fit<-bexpgeo_fit(Data.df,level = 0.95)
fit

}
\references{
Kozubowski, T.J., & Panorska, A.K. (2005). A Mixed bivariate distribution with exponential and geometric marginals. Journal of Statistical Planning and Inference, 134, 501-520.
\url{https://doi.org/10.1016/j.jspi.2004.04.010}
}
