% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg_distribution.R, R/bigg_distribution.R
\name{bgammageo_fit}
\alias{bgammageo_fit}
\title{Fits the bivariate gamma geometric distribution (BGG) to data.}
\usage{
bgammageo_fit(data, level = 0.95)

bgammageo_fit(data, level = 0.95)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BGG model.}

\item{level}{confidence level espressed between 0 and 1 (Default is 0.95).}

\item{data}{bivariate vector  (X,N) observations from BGG model.}

\item{level}{confidence level espressed between 0 and 1 (Default is 0.95).}
}
\value{
list of parameter estimates, confidence interval, deviance and covariance matrix.

list of parameter estimates, confidence interval, deviance and covariance matrix.
}
\description{
This function computes the parameter estimates, confidence interval, deviance and covariance matrix of gamma geometric distribution.
'
bgammageo_fit BGG model to data.

This function computes the parameter estimates, confidence interval, deviance and covariance matrix of bivariate inverse-Gaussian geometric distribution.
'
bgammageo_fit BGG model to data.
}
\examples{
Data.df<- rbgammageo(200,alpha=1.5, beta=2, p=0.6)
fit <- bgammageo_fit(Data.df)
fit

Data.df<- rbgammageo(200,alpha=1.5, beta=2, p=0.6)
fit <- bgammageo_fit(Data.df)
fit

}
\references{
Barreto-Souza, W. (2012). Bivariate gamma-geometric law and its induced Lévy process . Journal of Multivariate Analysis, 109:130-145.
\url{https://doi.org/10.1016/j.jmva.2012.03.004}

Barreto-Souza, W. and Silva R. B. (2019). A bivariate infinitely divisible law for modeling the magnitude and duration of monotone periods of log-returns . Statistica Neerlandica, 73:211-233.
\url{https://doi.org/10.1111/stan.12166}
}
