% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigg_distribution.R
\name{binvgaussgeo_fit}
\alias{binvgaussgeo_fit}
\title{Fits the bivariate inverse-Gaussian geometric distribution (BIGG) to data.}
\usage{
binvgaussgeo_fit(data, level = 0.95)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BIGG model.}

\item{level}{confidence level espressed between 0 and 1 (Default is 0.95).}
}
\value{
list of parameter estimates, confidence interval, deviance and covariance matrix.
}
\description{
This function computes the parameter estimates, confidence interval, deviance and covariance matrix of bivariate inverse-Gaussian geometric distribution.
'
binvgaussgeo_fit BIGG model to data.
}
\examples{
Data.df <- rbinvgaussgeo(50,mu=1.5, phi=2, p=0.6)
fit <- binvgaussgeo_fit(Data.df)
fit

}
\references{
Barreto-Souza, W. and Silva R. B. (2019). A bivariate infinitely divisible law for modeling the magnitude and duration of monotone periods of log-returns . Statistica Neerlandica, 73:211-233.
\url{https://doi.org/10.1111/stan.12166}
}
