% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{bmixexpgeo.init}
\alias{bmixexpgeo.init}
\title{parameter intialization for BMEG EM algoritm}
\usage{
bmixexpgeo.init(X, beta = NULL, q = NULL, k = 2)
}
\arguments{
\item{X}{data-vector X from BMEG distribution}

\item{beta}{scale paramters vectors for the mixture components, which must be numeric greater than 0}

\item{q}{is vector of probabilities for belong to the components, which must between 0 and 1 and the sum equal to 1}

\item{k}{number of mixture components}
}
\value{
list containg intial parameter estimates of beta and q
}
\description{
This function computes the parameter estimates of BMEG distribution using the EM algorithm.
}
\details{
Takes initial guess for the parameters and the algorithm will estmate the MLE.
}
\references{
Code adapted from Young et al (2017). mixtools Package : Tools for Analyzing Finite Mixture Models, R CRAN.
}
