% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{bmixexpgeo_em}
\alias{bmixexpgeo_em}
\title{EM algorithm function for the mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG)}
\usage{
bmixexpgeo_em(data, beta = NULL, q = NULL, k = 2, maxiter = 1000,
  tol = 1e-08, verb = FALSE)
}
\arguments{
\item{data}{dataframe of bivariate random vector (X,N) BMEG distribution.}

\item{beta}{scaled paramters vectors for the mixture components, which must be numeric greater than 0.}

\item{q}{is vector of probabilities for belong to the components, which must between 0 and 1 and the sum equal to 1.}

\item{k}{number of mixture components}

\item{maxiter}{maximum number of iterations.}

\item{tol}{tolerance value.}

\item{verb}{If TRUE, estimates are printed during each iteration.}
}
\value{
list containg parameter estimates, Deviance and data frame of iteration
}
\description{
This function computes the parameter estimates of BMEG distribution using the EM algorithm.
}
\details{
Takes initial guess for the parameters and the algorithm will estmate the MLE.
}
\examples{
data.df<-rbmixexpgeo(20, beta=c(1,2),p=0.6,q=c(0.4,0.6))
fit<-bmixexpgeo_em(data.df,k=2)
fit$par

}
\references{
Amponsah, C. K.,  Kozubowski, T. J. and Panorska (2019). A computational approach to estimation of discrete Pareto parameters. Inprint.
}
