% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixgnb_distribution.R
\name{bmixgammanbinom_fit}
\alias{bmixgammanbinom_fit}
\title{Fits the bivariate mixed gamma negative binomial distribution  (BMixGNB) to data.}
\usage{
bmixgammanbinom_fit(data, r = NULL, level = 0.95)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BMixGNB model.}

\item{r}{numeric parameter, which must NULL if the parameter in unknown or an integer.}

\item{level}{confidence level espressed between 0 and 1 (Default is 0.95).}
}
\value{
list of parameter estimates, confidence interval, deviance and covariance matrix.
}
\description{
This function computes the parameter estimates, confidence interval, log-likelihood and covariance matrix for the parameters in the bivariate mixed gamma negative binomial distribution
'
bmixgammanbinom_fit BMixGNB model to data.
}
\examples{
data.df<-rbmixgammanbinom(20,alpha=1.5, beta=2, r=3, p=0.6)
fit <- bmixgammanbinom_fit(data.df)
fit

}
\references{
Barreto-Souza, W. (2012). Bivariate gamma-geometric law and its induced Lévy process . Journal of Multivariate Analysis, 109:130-145.
\url{https://doi.org/10.1016/j.jmva.2012.03.004}
}
