% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg_distribution.R
\name{dbgammageo}
\alias{dbgammageo}
\title{The bivariate gamma geometric distribution (BGG)}
\usage{
dbgammageo(data, alpha, beta, p, log.p = FALSE)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BGG model.}

\item{alpha}{numeric parameter which must be  greater than  0.}

\item{beta}{numeric parameter which must be  greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of densities.
}
\description{
Density function for bivariate gamma geometric distribution with parameters \eqn{\alpha >0, \beta > 0} and p in (0,1).
}
\details{
dbgammageo is the density function for BGG model.
}
\examples{
data.df<-rbgammageo(20, alpha=1.5, beta=2, p=0.6)
den<-dbgammageo(data.df, alpha=1.5, beta=2, p=0.6)
den

}
\references{
Barreto-Souza, W. (2012). Bivariate gamma-geometric law and its induced Lévy process . Journal of Multivariate Analysis, 109:130-145.
\url{https://doi.org/10.1016/j.jmva.2012.03.004}
}
