% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigg_distribution.R
\name{dbinvgaussgeo}
\alias{dbinvgaussgeo}
\title{The bivariate inverse-Gaussian geometric distribution (BIGG)}
\usage{
dbinvgaussgeo(data, mu, phi, p, log.p = FALSE)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BGG model.}

\item{mu}{shape paramter which must be numeric greater than  0.}

\item{phi}{scale paramter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of densities.
}
\description{
Density function for bivariate inverse-Gaussian geometric distribution with parameters \eqn{\mu >0, \phi > 0} and p in (0,1).
}
\details{
dbinvgaussgeo is the density function for BGG model.
}
\examples{
data.df<-rbinvgaussgeo(20,mu=1.5, phi=2, p=0.6)
den<-rbinvgaussgeo(20,mu=1.5, phi=2, p=0.6)
den

}
\references{
Barreto-Souza, W. and Silva R. B., (2019). A bivariate infinitely divisible law for modeling the magnitude and duration of monotone periods of log-returns . Statistica Neerlandica, 73:211-233.
\url{https://doi.org/10.1111/stan.12166}
}
