% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blg_distribution.R
\name{dblomaxgeo}
\alias{dblomaxgeo}
\title{The bivariate distribution with Lomax and geometric margins (BLG)}
\usage{
dblomaxgeo(data, alpha, beta, p, log.p = FALSE)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BLG model.}

\item{alpha}{parameter which must be numeric greater than  0.}

\item{beta}{parameter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of densities.
}
\description{
Density function for bivariate distribution with Lomax and geometric margins (BLG) distribution with parameters \eqn{\alpha >0, \beta > 0} and p in [0,1].

[0,1]: R:0,1%5C
}
\details{
dblomaxgeo is the density function for BLG model.
}
\examples{
data.df<-rblomaxgeo(20, alpha=1.5, beta=2, p=0.6)
den<-dblomaxgeo(data.df, alpha=1.5, beta=2, p=0.6)
den

}
\references{
Arendarczyk, M. and Kozubowski T. J. and Panorska, A. k. (2018). A bivariate distribution with Lomax and geometric margins . Journal of the Korean Statistical Society, 47:405-422.
\url{https://doi.org/10.1016/j.jkss.2018.04.006}
}
