% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{dbmixexpgeo}
\alias{dbmixexpgeo}
\title{The mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG)}
\usage{
dbmixexpgeo(data, beta, p, q, log.p = FALSE)
}
\arguments{
\item{data}{is bivariate vector  (X,N) vector representing observations from BMEG model.}

\item{beta}{vector of scale paramters which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{q}{vector of numeric parameters each between 0 and 1, and sum equal to 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of densities.
}
\description{
Density function for bivariate distribution with mixture of exponential and geometric marginals (BMEG).
}
\details{
dbmixexpgeo is the density  function.
}
\examples{
data<-rbmixexpgeo(20, beta=c(1,2),p=0.6,q=c(0.4,0.6))
den<-dbmixexpgeo(data, beta=c(1,2),p=0.6,q=c(0.4,0.6))
den

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A Mixed bivariate distribution with mixture exponential and geometric marginals. Inprint.
}
