% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixgnb_distribution.R
\name{dbmixgammanbinom}
\alias{dbmixgammanbinom}
\title{The bivariate mixed gamma negative binomial distribution  (BMixGNB)}
\usage{
dbmixgammanbinom(data, alpha, beta, r, p, log.p = FALSE)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BMixGNB model.}

\item{alpha}{numeric paramter which must be numeric greater than  0.}

\item{beta}{numeric paramter which must be numeric greater than  0.}

\item{r}{integer paramter which must be greater than or equal to  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of densities.
}
\description{
Density function for bivariate mixed gamma negative binomial distribution  with parameters \eqn{\alpha >0, \beta > 0} and p in (0,1).
}
\details{
dbmixgammanbinom is the density function for BMixGNB model.
}
\examples{
data.df<-rmixgammanbinom(20,alpha=1.5, beta=2, r=3, p=0.6)
den<-dmixgammanbinom(data.df,alpha=1.5, beta=2, r=3, p=0.6)
den

}
\references{
Barreto-Souza, W. (2012). Bivariate gamma-geometric law and its induced Lévy process . Journal of Multivariate Analysis, 109:130-145.
\url{https://doi.org/10.1016/j.jmva.2012.03.004}
}
