% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{ddpareto}
\alias{ddpareto}
\title{The Discrete Pareto Distribution}
\usage{
ddpareto(N, delta, p, log.p = FALSE)
}
\arguments{
\item{N}{A vector of random sample from discrete Pareto.}

\item{delta}{A shape paramter which must be numeric greater than or equal to 0.}

\item{p}{A numeric parameter between 0 and 1}
}
\value{
A vector consiting of densities
}
\description{
Probaility mass function.
}
