% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exp_Tilted_Gamma.R
\name{detgamma}
\alias{detgamma}
\title{The Exponentially Tilted Gamma (ETG) distribution}
\usage{
detgamma(X, alpha, beta, K, log.p = FALSE)
}
\arguments{
\item{X}{A vector of random sample from ETG.}

\item{alpha}{A shape parameter which must be numeric greater than or equal to 0.}

\item{beta}{A scale parameter which must be numeric greater than or equal to 0.}

\item{K}{A numeric parameter between 0 and 1}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector representing densities \eqn{P(X=x)}.
}
\description{
Probability density function of ETG with parameters \eqn{\alpha,\beta \ge 0} and \eqn{\kappa \in [0,1]}.
}
\details{
detgamma gives the probability density(ies).
}
\examples{
prob <- detgamma(seq(0.1,5,0.1),alpha=0.2,beta=10, K=0.2)
prob

}
\references{
Amponsah, C. K. and  Kozubowski, T. J. (2021). A computational approach to estimation of discrete Pareto parameters. Preprint
}
