% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{dfmbeg}
\alias{dfmbeg}
\title{The finite mixture of bivariate distribution with mixture exponential and geometric (FMBEG) marginals.}
\usage{
dfmbeg(data, beta, p, q, pi, log.p = FALSE)
}
\arguments{
\item{data}{is bivariate vector  (X,N) vector representing observations from FMBEG model.}

\item{beta}{vector of parameters which must be numeric greater than  0.}

\item{p}{vector of numeric parameters between 0 and 1 of the finite mixture of geometric distribution.}

\item{q}{vector of gamma membership probabilities between 0 and 1, and sum equal to 1.}

\item{pi}{vector of geometric membership probabilities between 0 and 1, and sum equal to 1.}

\item{log.p}{logical; if TRUE, densities are given as logarithmic values.}
}
\value{
vector of densities.
}
\description{
Density function for FMBEG model.
}
\details{
dfmbeg is the density  function.
}
\examples{
data.df <- rfmbeg(10, beta = c(1,2,10), p=0.5, q=c(0.3,0.2,0.5), pi=1)
den <- dfmbeg(data=data.df, beta = c(1,2,10), p=0.5, q=c(0.3,0.2,0.5), pi=1)
den

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A finite mixture of bivariate distribution with mixture exponential and geometric (FMBEG) marginals. Inprint.
}
