% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMDP_distribution.R
\name{dgammamixdpareto}
\alias{dgammamixdpareto}
\title{The gamma mixture distcrete Pareto distribution (GMDP)}
\usage{
dgammamixdpareto(data, alpha, beta, delta, p, log.p = FALSE)
}
\arguments{
\item{data}{vector  (X,N) representing observation from GMDP model.}

\item{alpha}{numeric paramter which must be greater than  0.}

\item{beta}{numeric paramter which must be greater than  0.}

\item{delta}{numeric paramter which must be greater than or equal to  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of densities.
}
\description{
Density function for gamma mixture distcrete Pareto distribution with parameters \eqn{\alpha >0, \beta > 0, \delta \ge 0} and p in (0,1).
}
\details{
dgammamixdpareto is the density function of GMDP model.
}
\examples{
data.df<-rgammamixdpareto(20,alpha=1.5, beta=2, delta=0.1, p=0.3)
den<- dgammamixdpareto(data.df,alpha=1.5, beta=2, delta=0.1, p=0.3)
den

}
\references{
Amponsah, C. K., Kozubowski, T. J. and Panorska, A. K. (2019). A bivariate distribution with gamma mixture discrete Pareto marginals . In print.
\url{https://scholarworks.unr.edu/bitstream/handle/11714/2065/Amponsah_unr_0139M_12378.pdf?sequence=1&isAllowed=y}
}
