% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopois_ht.R
\name{dgeopois}
\alias{dgeopois}
\title{The Geometric with Poisson hidden truncation distribution}
\usage{
dgeopois(N, lambda, p, log.p = FALSE)
}
\arguments{
\item{N}{A vector of random sample from discrete Pareto.}

\item{lambda}{numeric parameter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector representing probabilities \eqn{P(N=n)}.
}
\description{
Probability mass function for geometric with Poisson hidden truncation distribution: parameters \eqn{\lambda > 0} and \eqn{p \in (0,1)}.
}
\details{
dgeopois gives the probability mass.
}
\examples{
prob<-dgeopois(seq(1,10,1), lambda=0.2, p=0.6)
prob

}
\references{
Amponsah, C. K. and  Kozubowski, T. J. (2020). Geometric with Poisson hidden truncation distribution.
\url{Yet to publish}
}
