% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgeo_distribution.R
\name{dmgeo}
\alias{dmgeo}
\title{The multivariate geometric distribution (MGEO)}
\usage{
dmgeo(data, theta, prob, log.p = FALSE)
}
\arguments{
\item{data}{is a data frame  MGEO model.}

\item{theta}{numeric parameter which must be greater than -1 but less than (1-min(p))/(1+min(p)).}

\item{prob}{vector of parameters between 0 and 1.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of probabilities.
}
\description{
Probability mass function for multivariate geometric distribution with parameters $p_1,...,p_d$ in (0,1)$ and \eqn{-1<\theta < \dfrac{1-\min(p_1,...,p_d)}{1+\min(p_1,...,p_d)}}.
}
\details{
dmgeo is the probability mass function.
}
\examples{
data<-rmgeo(n=20, theta = 2, prob = c(0.6, 0.1))
den<-dmgeo(data, theta=2, prob = c(0.6, 0.1))
den

}
\references{
Kozubowski, T.J., & Panorska, A.K. (2005). A Mixed bivariate distribution with exponential and geometric marginals. Journal of Statistical Planning and Inference, 134, 501-520.
\url{https://doi.org/10.1016/j.jspi.2004.04.010}
}
