% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dp_EM.R
\name{dpareto_em}
\alias{dpareto_em}
\title{EM algorithm function for discrete Pareto distribution}
\usage{
dpareto_em(data, delta = 1, p = 0.5, maxiter = 500, tol = 1e-16)
}
\arguments{
\item{data}{A vector of random sample from discrete pareto distribution}

\item{delta}{A shape paramter which must be numeric greater than or equal to 0}

\item{p}{A numeric parameter between 0 and 1}

\item{maxiter}{Maximum number of iterations}

\item{tol}{tolerance value}
}
\value{
A list containg parameter estimates, Deviance and data frame of iteration
}
\description{
Takes initial guess of parameters in discrete pareto and use Em algorithm to estmate the MLE
}
