% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{dpareto_fsa}
\alias{dpareto_fsa}
\title{Fisher Scoring algorithm (FSA) for estimation of discrete Pareto distribution parameters.}
\usage{
dpareto_fsa(
  N,
  delta = 1,
  p = NULL,
  lambda = NULL,
  maxiter = 20,
  tol = 1e-08,
  verb = FALSE
)
}
\arguments{
\item{N}{vector of random sample from discrete Pareto distribution}

\item{delta}{shape parameter which must be numeric greater than or equal to 0}

\item{p}{numeric parameter between 0 and 1}

\item{lambda}{learning parameter between 0 and 1}

\item{maxiter}{maximum number of iterations}

\item{tol}{tolerance value}

\item{verb}{If TRUE, estimates are printed during each iteration.}
}
\value{
list containg parameter estimates, Deviance and data frame of iteration
}
\description{
This function computes MLE of discrete Pareto distribution using the FSA algorithm.
}
\details{
Takes initial guess for the parameters in discrete Pareto and the algorithm will estimate the MLE.
}
\examples{
N<-rdpareto(500, delta=0.2,p=0.6)
fit<-dpareto_em(N,maxiter = 1000)
fit$par

}
\references{
Amponsah, C. K.,  Kozubowski, T. J. and Panorska (2019). A computational approach to estimation of discrete Pareto parameters. Inprint.
}
