% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{expPareto_test}
\alias{expPareto_test}
\title{Test of exponential versus Pareto type II (Lomax) distributions}
\usage{
expPareto_test(X)
}
\arguments{
\item{X}{data-vector.}
}
\value{
vector of samples generate from BLG distribution.
}
\description{
Performs \eqn{H_0:\alpha =0 \quad vs \quad H_0:\alpha =0}.  \eqn{\alpha =0} corresponds to exponetial distribution and \eqn{\alpha >0} corresponds to Pareto type II distribution.
}
\details{
expPareto_test is the function for testing exponential vs. Pareto type II distrbutions
}
\examples{
X<- Renext:: rlomax(100,scale = 0.200,shape = 0.5)
#X<-rexp(300, rate = 1/10)
test<- expPareto_test
test

}
\references{
Arendarczyk, M. and Kozubowski T. J. and Panorska, A. k. (2018). A bivariate distribution with Lomax and geometric margins . Journal of the Korean Statistical Society, 47:405-422.
\url{https://doi.org/10.1016/j.jkss.2018.04.006}
}
