% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{fmbeg_em}
\alias{fmbeg_em}
\title{EM algorithm function for the mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG)}
\usage{
fmbeg_em(data, beta = NULL, p = NULL, q = NULL, pi = NULL, m = 2,
  l = 1, maxiter = 1000, tol = 1e-08, verb = FALSE)
}
\arguments{
\item{data}{dataframe of bivariate random vector (X,N) BMEG distribution.}

\item{beta}{numeric parameter vectors, which must be numeric greater than 0. Default value is NULL.}

\item{p}{vector of numeric parameters between 0 and 1. Default value is NULL.}

\item{q}{vector of gamma membership probabilities between 0 and 1, and sum equal to 1. Default value is NULL.}

\item{pi}{vector of geometric membership probabilities between 0 and 1, and sum equal to 1. Default value is NULL.}

\item{m}{number of gamma components. Default value is 2.}

\item{l}{number of geometric components. Default value is 1.}

\item{maxiter}{maximum number of iterations. Default value is 1000.}

\item{tol}{tolerance value. Default value is 1e-08}

\item{verb}{If TRUE, estimates are printed during each iteration.}
}
\value{
list containing data frame parameter estimates, log-likelihood value and data frame of iteration
}
\description{
This function computes the parameter estimates of FMBEG distribution using the EM algorithm.
}
\details{
Takes initial guess for the parameters and the algorithm will estimate the MLE.
}
\examples{
data.df <- rfmbeg(100, beta = c(1,2,10), p=c(0.5,0.8,0.01), q=c(0.3,0.2,0.5), pi=c(0.2,0.3,0.5))
fit <- fmbeg_em(data=data.df, m=3,l=3)
fit$par

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A Mixed bivariate distribution with mixture exponential and geometric marginals. Inprint.
}
