% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{fmbeg_gamma.init}
\alias{fmbeg_gamma.init}
\title{parameter initialization for BMEG EM algorithm}
\usage{
fmbeg_gamma.init(data, beta = NULL, q = NULL, m)
}
\arguments{
\item{data}{data-vector from FMBEG distribution}

\item{beta}{numeric parameters, which must be numeric greater than 0. Default value is NULL}

\item{q}{vector of gamma membership probabilities between 0 and 1, and sum equal to 1. Default value is NULL.}

\item{m}{number of gamma components. Default value is 2.}
}
\value{
list containing initial parameter estimates of beta and q
}
\description{
This function computes the parameter estimates of BMEG distribution using the EM algorithm.
}
\details{
Takes initial guess for the parameters and the algorithm will estimate the MLE.
}
\references{
Code adapted from Young et al. (2017). mixtools Package : Tools for Analyzing Finite Mixture Models, R CRAN.
}
