% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{fmbeg_geo.init}
\alias{fmbeg_geo.init}
\title{parameter initialization for BMEG EM algorithm}
\usage{
fmbeg_geo.init(data, p = NULL, pi = NULL, l)
}
\arguments{
\item{data}{data from FMBEG distribution}

\item{p}{vector of numeric parameters between 0 and 1. Default value is NULL.}

\item{pi}{vector of geometric membership probabilities between 0 and 1, and sum equal to 1. Default value is NULL.}

\item{l}{number of geometric components. Default value is 1.}
}
\value{
list containing initial parameter estimates of p and membership probabilities
}
\description{
This function computes the parameter estimates of BMEG distribution using the EM algorithm.
}
\details{
Takes initial guess for the parameters and the algorithm will estimate the MLE.
}
\references{
Code adapted from Young et al. (2017). mixtools Package : Tools for Analyzing Finite Mixture Models, R CRAN.
}
