% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{func_mgamma_em}
\alias{func_mgamma_em}
\title{E and M steps for mixture of gamma distributions part of FMBEG}
\usage{
func_mgamma_em(data, beta, q, m)
}
\arguments{
\item{data}{data-vector from FMBEG distribution.}

\item{beta}{numeric parameters, which must be numeric greater than 0.}

\item{q}{vector of gamma membership probabilities between 0 and 1, and sum equal to 1}

\item{m}{number of gamma components.}
}
\value{
list containing  parameter estimates of p and pi
}
\description{
This function computes the beta and q of FMBEG distribution.
}
