% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMDP_estimation.R
\name{gammamixdpareto_fit}
\alias{gammamixdpareto_fit}
\title{Fits the gamma mixture distcrete Pareto distribution (GMDP) to data.}
\usage{
gammamixdpareto_fit(data, delta = 1, p = 0.5, method = "EM")
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from GMDP model.}

\item{delta}{initial guess of true parameter \eqn{\delta} which is numeric and must be greater than  0 (Default value is 1).}

\item{p}{nitial guess of true parameter p which must be numeric value between 0 and 1 (Default value is 0.5).}

\item{method}{method of estimation: \eqn{EM=}EM algorithm or \eqn{MLE=}maximum likelihood estimation (Default method is EM).}
}
\value{
vector of parameter estimates.
}
\description{
This function computes the estimates  of parameter in GMDP model.
}
\details{
bgammageo_fit GMDP model to data.
}
\examples{
Data.df<- rgammamixdpareto(n=1000,alpha=1,beta=2,delta=0.45, p=0.8)
fit <- gammamixdpareto_fit(data = Data.df)
fit

}
\references{
Amponsah, C. K., Kozubowski, T. J. and Panorska, A. K. (2019). A bivariate distribution with gamma mixture discrete Pareto marginals . In print.
\url{https://scholarworks.unr.edu/bitstream/handle/11714/2065/Amponsah_unr_0139M_12378.pdf?sequence=1&isAllowed=y}
}
