% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgeo_distribution.R
\name{mgeo_fit}
\alias{mgeo_fit}
\title{Fits the multivariate geometric distribution (MGEO)}
\usage{
mgeo_fit(data)
}
\arguments{
\item{data}{is data.frame of observations from MGEO model}
}
\value{
list of parameter estimates and deviance
}
\description{
This function computes the parameter estimates of MGEO .
}
\details{
mgeo_fit  fits MGEO model to data.
}
\examples{
Data.df<- rmgeo(n = 100,theta = 10, prob= c(0.23, 0.7))
fit<-mgeo_fit(Data.df)
fit

}
\references{
Kozubowski, T.J., & Panorska, A.K. (2005). A Mixed bivariate distribution with exponential and geometric marginals. Journal of Statistical Planning and Inference, 134, 501-520.
\url{https://doi.org/10.1016/j.jspi.2004.04.010}
}
