% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg_distribution.R
\name{pbgammageo}
\alias{pbgammageo}
\title{The bivariate gamma geometric distribution (BGG)}
\usage{
pbgammageo(data, alpha, beta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BGG model.}

\item{alpha}{numeric paramter which must be  greater than  0.}

\item{beta}{numeric paramter which must be  greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x, N \leq n]}, otherwise, \eqn{P[X > x, N > n]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of distribution.
}
\description{
Distribution function for bivariate gamma geometric distribution with parameters \eqn{\alpha >0, \beta > 0} and p in (0,1).
}
\details{
pbgammageo is the distribution function for BGG model.
}
\examples{
data.df <- rbgammageo(20, alpha=1.5, beta=2, p=0.6)
prob <- pbgammageo(data.df, alpha=1.5, beta=2, p=0.6)
prob

}
\references{
Barreto-Souza, W. (2012). Bivariate gamma-geometric law and its induced Lévy process . Journal of Multivariate Analysis, 109:130-145.
\url{https://doi.org/10.1016/j.jmva.2012.03.004}
}
