% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigg_distribution.R
\name{pbinvgaussgeo}
\alias{pbinvgaussgeo}
\title{The bivariate inverse-Gaussian geometric distribution (BIGG)}
\usage{
pbinvgaussgeo(data, mu, phi, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{data}{bivariate vector  (X,N) observations from BGG model.}

\item{mu}{shape paramter which must be numeric greater than  0.}

\item{phi}{scale paramter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x, N \leq n]}, otherwise, \eqn{P[X > x, N > n]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of distribution.
}
\description{
Distribution function for bivariate inverse-Gaussian geometric distribution with parameters \eqn{\mu >0, \phi > 0} and p in (0,1).
}
\details{
pbgammageo is the distribution function for BGG model.
}
\examples{
data.df<-rbgammageo(20, alpha=1.5, beta=2, p=0.6)
den<-pbgammageo(data.df, alpha=1.5, beta=2, p=0.6)
den

}
\references{
Barreto-Souza, W. and Silva R. B. (2019). A bivariate infinitely divisible law for modeling the magnitude and duration of monotone periods of log-returns . Statistica Neerlandica, 73:211-233.
\url{https://doi.org/10.1111/stan.12166}
}
