% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{pbmixexpgeo}
\alias{pbmixexpgeo}
\title{The mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG)}
\usage{
pbmixexpgeo(data, beta, p, q, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{data}{is bivariate vector  (X,N) vector representing observations from BMEG model.}

\item{beta}{vector of scale paramters which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{q}{vector of numeric parameters each between 0 and 1, and sum equal to 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x, N \leq n]}, otherwise, \eqn{P[X > x, N > n]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of distribution.
}
\description{
Distribution function for the mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG).
}
\details{
pbmixexpgeo is the distribution  function.
}
\examples{
data<-rbmixexpgeo(20, beta=c(1,2),p=0.6,q=c(0.4,0.6))
prob<-pbmixexpgeo(data, beta=c(1,2),p=0.6,q=c(0.4,0.6))
prob

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A Mixed bivariate distribution with mixture exponential and geometric marginals. Inprint.
}
