% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{pdpareto}
\alias{pdpareto}
\title{The Discrete Pareto Distribution}
\usage{
pdpareto(q, delta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{A vector of quantiles values from discrete pareto distribution.}

\item{delta}{A shape paramter which must be numeric greater than or equal to 0.}

\item{p}{A numeric parameter between 0 and 1}

\item{lower.tail}{cummulative probabilities if TRUE otherwise survival probilities}

\item{log.p}{log probabilities if TRUE}
}
\value{
A vector cumulative probabilities
}
\description{
Distribution function.
}
