% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{pfmbeg}
\alias{pfmbeg}
\title{The finite mixture of bivariate distribution with mixture exponential and geometric (FMBEG) marginals.}
\usage{
pfmbeg(data, beta, p, q, pi, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{data}{is bivariate vector  (X,N) vector representing observations from BMEG model.}

\item{beta}{vector of parameters which must be numeric greater than  0.}

\item{p}{vector of numeric parameters between 0 and 1 of the finite mixture of geometric distribution.}

\item{q}{vector of gamma membership probabilities between 0 and 1, and sum equal to 1.}

\item{pi}{vector of geometric membership probabilities between 0 and 1, and sum equal to 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x, N \leq n]}, otherwise, \eqn{P[X > x, N > n]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of distribution.
}
\description{
Distribution function for FMBEG.
}
\details{
pfmbeg is the distribution  function.
}
\examples{
data.df <- rfmbeg(10, beta = c(1,2,10), p=0.5, q=c(0.3,0.2,0.5), pi=1)
prob <- pfmbeg(data=data.df, beta = c(1,2,10), p=0.5, q=c(0.3,0.2,0.5), pi=1)
prob

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A Mixed bivariate distribution with mixture exponential and geometric marginals. Inprint.
}
