% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopois_ht.R
\name{pgeopois}
\alias{pgeopois}
\title{The Geometric with Poisson hidden truncation distribution}
\usage{
pgeopois(q, lambda, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles representing numbers from discrete Pareto distribution.}

\item{lambda}{numeric parameter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[N\leq n]}, otherwise, \eqn{$P[N> n]$}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A vector cumulative probabilities
}
\description{
Distribution function for Geometric with Poisson hidden truncation distribution, where parameters: \eqn{\lambda > 0} and \eqn{p \in (0,1)}.
}
\details{
pgeopois gives the distribution function.
}
\examples{
prob<-pgeopois(seq(1,10, 1), lambda=0.2, p=0.6)
prob

}
\references{
Amponsah, C. K. and  Kozubowski, T. J. (2020). Geometric with Poisson hidden truncation distribution.
\url{Yet to publish}
}
