% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{qdpareto}
\alias{qdpareto}
\title{The Discrete Pareto Distribution}
\usage{
qdpareto(prob, delta, p)
}
\arguments{
\item{prob}{Vector of probabilities.}

\item{delta}{A shape paramter which must be numeric greater than or equal to 0.}

\item{p}{A numeric parameter between 0 and 1}
}
\value{
A vector quantiles
}
\description{
Quantile function.
}
