% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigg_distribution.R
\name{rbinvgaussgeo}
\alias{rbinvgaussgeo}
\title{The bivariate inverse-Gaussian geometric distribution (BIGG)}
\usage{
rbinvgaussgeo(n, mu, phi, p)
}
\arguments{
\item{n}{size of sample.}

\item{mu}{shape paramter which must be numeric greater than  0.}

\item{phi}{scale paramter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector of samples generate from BGG distribution.
}
\description{
Random sample generating function for bivariate inverse-Gaussian geometric distribution with parameters \eqn{\mu >0, \phi > 0} and p in (0,1).
}
\details{
rbinvgaussgeo generates random samples from BIGG distribution.
}
\examples{
data.df<-rbinvgaussgeo(20,mu=1.5, phi=2, p=0.6)
data.df

}
\references{
Barreto-Souza, W. and Silva R. B. (2019). A bivariate infinitely divisible law for modeling the magnitude and duration of monotone periods of log-returns . Statistica Neerlandica, 73:211-233.
\url{https://doi.org/10.1111/stan.12166}
}
