% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blg_distribution.R
\name{rblomaxgeo}
\alias{rblomaxgeo}
\title{The bivariate distribution with Lomax and geometric margins (BLG)}
\usage{
rblomaxgeo(n, alpha, beta, p)
}
\arguments{
\item{n}{size of sample.}

\item{alpha}{parameter which must be numeric greater than  0.}

\item{beta}{parameter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector of samples generate from BLG distribution.
}
\description{
Random sample generating function for bivariate distribution with Lomax and geometric margins (BLG) distribution with parameters \eqn{\alpha >0, \beta > 0} and p in [0,1].

[0,1]: R:0,1%5C
}
\details{
rblomaxgeo generates random samples from BLG distribution.
}
\examples{
data.df<-rblomaxgeo(20,alpha=1.5, beta=2, p=0.6)
data.df

}
\references{
Arendarczyk, M. and Kozubowski T. J. and Panorska, A. k. (2018). A bivariate distribution with Lomax and geometric margins . Journal of the Korean Statistical Society, 47:405-422.
\url{https://doi.org/10.1016/j.jkss.2018.04.006}
}
