% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{rbmixexpgeo}
\alias{rbmixexpgeo}
\title{The mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG)}
\usage{
rbmixexpgeo(n, beta, p, q)
}
\arguments{
\item{n}{size of sample.}

\item{beta}{vector of scale parameters which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{q}{vector of numeric parameters each between 0 and 1, and sum equal to 1.}
}
\value{
vector of random samples generate from BMEG model.
}
\description{
Random sample generating function for the mixed bivariate distribution with mixture of exponential and geometric marginals (BMEG).
}
\details{
rbexpgeo generates random sample from BEG distribution.
}
\examples{
N<-rbmixexpgeo(20, beta=c(1,2),p=0.6,q=c(0.4,0.6))
N

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A Mixed bivariate distribution with mixture exponential and geometric marginals. Inprint.
}
