% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmixgnb_distribution.R
\name{rbmixgammanbinom}
\alias{rbmixgammanbinom}
\title{The bivariate mixed gamma negative binomial distribution (BMixGNB)}
\usage{
rbmixgammanbinom(n, alpha, beta, r, p)
}
\arguments{
\item{n}{size of sample.}

\item{alpha}{numeric paramter which must be numeric greater than  0.}

\item{beta}{numeric paramter which must be numeric greater than  0.}

\item{r}{integer paramter which must be greater than or equal to  0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector of samples generate from BGG distribution.
}
\description{
Random sample generating function for bivariate mixed gamma negative binomial distribution with parameters \eqn{\alpha >0, \beta > 0} and p in (0,1).
}
\details{
rbmixgammanbinom generates random samples from BMixGNB distribution.
}
\examples{
data.df<-rbmixgammanbinom(20,alpha=1.5, beta=2, r=3, p=0.6)
data.df

}
\references{
Barreto-Souza, W. (2012). Bivariate gamma-geometric law and its induced Lévy process . Journal of Multivariate Analysis, 109:130-145.
\url{https://doi.org/10.1016/j.jmva.2012.03.004}
}
