% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discrete_HT_models.R
\name{rdHTgamma}
\alias{rdHTgamma}
\title{The family of discrete distributions with gamma hidden truncation.}
\usage{
rdHTgamma(n, x, alpha, beta, distr, lambda = NULL, prob = NULL,
  delta = NULL, size = NULL)
}
\arguments{
\item{n}{size of sample.}

\item{x}{numeric truncation value which must be greater than or equal to  0.}

\item{alpha}{numeric gamma shape parameter which must be greater than  0.}

\item{beta}{numeric gamma rate parameter which must be greater than  0.}

\item{distr}{a description of the discrete distribution which is truncated by gamma: "Poisson", "Geometric", "DPareto"(for discrete Pareto) and "nbinom" (for Negative binomial).}

\item{lambda}{numeric Poisson parameter which must be greater than 0.}

\item{prob}{numeric parameter between 0 and 1. Must be specified when distribution is either geometric or discrete Pareto or negative binomial.}

\item{delta}{numeric discrete Pareto shape parameter which must be greater than or equal to 0.}

\item{size}{numeric negative binomial parameter which must be greater than  0.}
}
\value{
vector of samples generate from discrete Pareto distribution.
}
\description{
Random sample generating function for the family of discrete distributions with gamma hidden truncation: parameters \eqn{\alpha,\beta,\lambda,\delta, \mu \ge 0} and \eqn{p,\theta\in (0,1).} p in (0,1).
}
\details{
rdHTgamma generates random sample from discrete Pareto distribution.
}
\examples{
N <- rdHTgamma(n=2, x=0.05, alpha = 1,beta = 100, distr = "Geometric",p = 0.5)
N

}
\references{
Amponsah, C. K. and  Kozubowski, T. J. (2021). Discrete Hidden Truncation models. In preparation.
\url{https://doi.org/not-done}
}
