% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{rdpareto}
\alias{rdpareto}
\title{The Discrete Pareto Distribution}
\usage{
rdpareto(n, delta, p)
}
\arguments{
\item{n}{Size of sample.}

\item{delta}{A shape paramter which must be numeric greater than or equal to 0.}

\item{p}{A numeric parameter between 0 and 1}
}
\value{
A vector of sample generate from discrete Pareto distribution.
}
\description{
Random sample generation.
}
