% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbeg_EM.R
\name{rfmbeg}
\alias{rfmbeg}
\title{The finite mixture of bivariate distribution with mixture exponential and geometric (FMBEG) marginals.}
\usage{
rfmbeg(n, beta, p, q, pi)
}
\arguments{
\item{n}{size of sample.}

\item{beta}{vector of parameters which must be numeric greater than  0.}

\item{p}{vector of numeric parameters between 0 and 1 of the finite mixture of geometric distribution.}

\item{q}{vector of gamma membership probabilities between 0 and 1, and sum equal to 1.}

\item{pi}{vector of geometric membership probabilities between 0 and 1, and sum equal to 1.}
}
\value{
vector of random samples generate from BMEG model.
}
\description{
Random sample generating function for the FMBEG model.
}
\details{
rfmbeg generates random sample from FMBEG distribution.
}
\examples{
data.df <- rfmbeg(10, beta = c(1,2,10), p=0.5, q=c(0.3,0.2,0.5), pi=1)
data.df

}
\references{
Amponsah, C. K. and Kozubowski, T.J., and Panorska, A.K. (2020). A finite mixture of bivariate distribution with mixture exponential and geometric (FMBEG) marginals. Inprint.
}
