% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMDP_distribution.R
\name{rgammamixdpareto}
\alias{rgammamixdpareto}
\title{The gamma mixture discrete Pareto distribution (GMDP)}
\usage{
rgammamixdpareto(n, alpha, beta, delta, p)
}
\arguments{
\item{n}{size of sample.}

\item{alpha}{numeric parameter which must be greater than  0.}

\item{beta}{numeric parameter which must be greater than  0.}

\item{delta}{numeric parameter which must be greater than or equal to  0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector of random samples from GMDP distribution.
}
\description{
Random sample generating function for gamma mixture discrete Pareto distribution with parameters \eqn{\alpha >0, \beta > 0, \delta \ge 0} and p in (0,1).
}
\details{
rgammamixdpareto generates random samples from GMDP distribution.
}
\examples{
data.df<-rgammamixdpareto(20,alpha=1.5, beta=2, delta=0.1, p=0.3)
data.df

}
\references{
Amponsah, C. K., Kozubowski, T. J. and Panorska, A. K. (2019). A bivariate distribution with gamma mixture discrete Pareto marginals . In print.
\url{https://scholarworks.unr.edu/bitstream/handle/11714/2065/Amponsah_unr_0139M_12378.pdf?sequence=1&isAllowed=y}
}
