% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgeo_distribution.R
\name{rmgeo}
\alias{rmgeo}
\title{The multivariate geometric distribution (MGEO)}
\usage{
rmgeo(n, theta, prob)
}
\arguments{
\item{n}{size of sample.}

\item{theta}{numeric parameter which must be greater than -1 but less than (1-min(p))/(1+min(p)).}

\item{prob}{vector of parameters between 0 and 1.}
}
\value{
vector of random samples generate from MGEO model.
}
\description{
Random sample generating function for multivariate geometric distribution with parameters $p_1,...,p_d$ in (0,1)$ and \eqn{-1<\theta < \dfrac{1-\min(p_1,...,p_d)}{1+\min(p_1,...,p_d)}}.
}
\details{
rmgeo generates random sample from MGEO distribution.
}
\examples{
N<-rmgeo(n=20, theta = 0.6, prob = c(0.2, 0.5))
N

}
\references{
Amponsah, C. K., & Kozubowski, T.J. (2022). A mixed multivariate distribution with bivariate exponential and geometric marginals. Inprint.
\url{https://doi.org/10.1016/j.jspi.2004.04.010}
}
