% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{ddpareto}
\alias{ddpareto}
\title{The discrete Pareto distribution}
\usage{
ddpareto(N, delta, p, log.p = FALSE)
}
\arguments{
\item{N}{A vector of random sample from discrete Pareto.}

\item{delta}{A shape parameter which must be numeric greater than or equal to 0.}

\item{p}{A numeric parameter between 0 and 1}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector representing probabilities \eqn{P(N=n)}.
}
\description{
Probability mass function for discrete Pareto distribution with parameters \eqn{\delta \ge 0} and p in (0,1).
}
\details{
ddpareto gives the probability mass.
}
\examples{
prob<-ddpareto(seq(1,10,1),delta=0.2,p=0.6)
prob

}
\references{
Buddana, A. and  Kozubowski, T. J. (2014). Discrete Pareto distribution. Journal of Economics and Quality Control, 29(2):143-156.
\url{https://doi.org/10.1515/eqc-2014-0014}
}
