% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dp_EM.R
\name{dpareto_em}
\alias{dpareto_em}
\title{EM algorithm function for discrete Pareto distribution}
\usage{
dpareto_em(data, delta = 1, p = 0.5, maxiter = 500, tol = 1e-16)
}
\arguments{
\item{data}{vector of random sample from discrete pareto distribution}

\item{delta}{shape paramter which must be numeric greater than or equal to 0}

\item{p}{numeric parameter between 0 and 1}

\item{maxiter}{maximum number of iterations}

\item{tol}{tolerance value}
}
\value{
list containg parameter estimates, Deviance and data frame of iteration
}
\description{
This function computes the parameter estimates of discrete Pareto distribution using the EM algorithm.
}
\details{
Takes initial guess for the parameters in discrete pareto and the algorithm will estmate the MLE.
}
\examples{
N<-rdpareto(500, delta=0.2,p=0.6)
fit<-dpareto_em(N,maxiter = 1000)
fit$par

}
\references{
Amponsah, C. K.,  Kozubowski, T. J. and Panorska (2019). A computational approach to estimation of discrete Pareto parameters. Inprint.
}
