% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_mixed_distribution.R
\name{mgammageo_fit}
\alias{mgammageo_fit}
\title{Fits the multivariate mixed gamma-geometric distributions (MMGG)}
\usage{
mgammageo_fit(data)
}
\arguments{
\item{data}{is data.frame of observations, (X, N, Y, M)}
}
\value{
list of parameter estimates and the log-likelihood value
}
\description{
This function computes the parameters of the model .
}
\details{
mgammageo_fit  fits MMGG model to data.
}
\examples{
Data.df<- data.df<- rmgammageo(20, alpha=c(0.5, 1.5), beta=c(20, 50), prob=c(0.6, 0.4), theta= -0.2)
fit <- mgammageo_fit(Data.df)
fit

}
\references{
Amponsah, C. K. and Kozubowski, T.J., (2022). Inprint
}
