% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beg_distribution.R
\name{pbexpgeo}
\alias{pbexpgeo}
\title{The bivariate exponential geometric distrubution (BEG)}
\usage{
pbexpgeo(data, beta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{data}{is bivariate vector  (X,N) vector representing observations from BEG model.}

\item{beta}{numeric paramter which must be numeric greater than 0.}

\item{p}{numeric parameter between 0 and 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x, N \leq n]}, otherwise, \eqn{P[X > x, N > n]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
vector of distribution.
}
\description{
Distribution function for bivariate exponential geometric distribution with parameters \eqn{\beta > 0} and p in (0,1).
}
\details{
pbexpgeo is the distribution  function.
}
\examples{
data<-rbexpgeo(20, beta=2,p=0.6)
den<-pbexpgeo(data, beta=2,p=0.6)
den

}
\references{
Kozubowski, T.J., & Panorska, A.K. (2005). A Mixed bivariate distribution with exponential and geometric marginals. Journal of Statistical Planning and Inference, 134, 501-520.
\url{https://doi.org/10.1016/j.jspi.2004.04.010}
}
