% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blg_distribution.R
\name{pblomaxgeo}
\alias{pblomaxgeo}
\title{The bivariate distribution with Lomax and geometric margins (BLG)}
\usage{
pblomaxgeo(q, alpha, beta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{alpha}{paramter which must be numeric greater than  0.}

\item{beta}{paramter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x, N \leq n]}, otherwise, \eqn{P[X > x, N > n]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{data}{bivariate vector  (X,N) observations from BLG model.}
}
\value{
vector of distribution.
}
\description{
Density function for bivariate distribution with Lomax and geometric margins (BLG) distribution with parameters \eqn{\alpha >0, \beta > 0} and p in \link{0,1}.
}
\details{
pblomaxgeo is the distribution function for BLG model.
}
\examples{
data.df<-rblomaxgeo(20, alpha=1.5, beta=2, p=0.6)
den<-pblomaxgeo(data.df, alpha=1.5, beta=2, p=0.6)
den

}
\references{
Barreto-Souza, W. (2012). A bivariate distribution with Lomax and geometric margins . Journal of the Korean Statistical Society, 47:405-422.
\url{https://doi.org/10.1016/j.jkss.2018.04.006}
}
