% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{pdpareto}
\alias{pdpareto}
\title{The discrete Pareto distribution}
\usage{
pdpareto(q, delta, p, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles representing numbers from discrete Pareto distribution.}

\item{delta}{shape parameter which must be numeric greater than or equal to 0.}

\item{p}{numeric parameter between 0 and 1.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[N\leq n]}, otherwise, \eqn{$P[N> n]$}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A vector cumulative probabilities
}
\description{
Distribution function for discrete Pareto distribution with parameters \eqn{\delta \ge 0} and p in (0,1).
}
\details{
pdpareto gives the distribution function.
}
\examples{
prob<-pdpareto(seq(1,10, 1),delta=0.2,p=0.6)
prob

}
\references{
Buddana, A. and  Kozubowski, T. J. (2014). Discrete Pareto distribution. Journal of Economics and Quality Control, 29(2):143-156.
\url{https://doi.org/10.1515/eqc-2014-0014}
}
