% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_pareto.R
\name{qdpareto}
\alias{qdpareto}
\title{The discrete Pareto distribution}
\usage{
qdpareto(prob, delta, p)
}
\arguments{
\item{prob}{Vector of probabilities.}

\item{delta}{shape paramter which must be numeric greater than or equal to 0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector quantiles from discrete pareto distribution.
}
\description{
Quantile function for discrete Pareto distribution with parameters \eqn{\delta \ge 0} and p in (0,1).
}
\details{
qdpareto gives the quantile function.
}
\examples{
q<-rdpareto(seq(0.1,0.6,0.1), delta=0.2,p=0.6)
q

}
\references{
Buddana, A. and  Kozubowski, T. J. (2014). Discrete Pareto distribution. Journal of Economics and Quality Control, 29(2):143-156.
\url{https://doi.org/10.1515/eqc-2014-0014}
}
