% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beg_distribution.R
\name{rbexpgeo}
\alias{rbexpgeo}
\title{The bivariate exponential geometric distrubution (BEG)}
\usage{
rbexpgeo(n, beta, p)
}
\arguments{
\item{n}{size of sample.}

\item{beta}{numeric paramter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector of random samples generate from BEG model.
}
\description{
Random sample generating function for bivariate exponential geometric distribution with parameters \eqn{\beta > 0} and p in (0,1).
}
\details{
rbexpgeo generates random sample from BEG distribution.
}
\examples{
N<-rbexpgeo(20, beta=2,p=0.6)
N

}
\references{
Kozubowski, T.J., & Panorska, A.K. (2005). A Mixed bivariate distribution with exponential and geometric marginals. Journal of Statistical Planning and Inference, 134, 501-520.
\url{https://doi.org/10.1016/j.jspi.2004.04.010}
}
