% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopois_ht.R
\name{rgeopois}
\alias{rgeopois}
\title{The Geometric with Poisson hidden truncation distribution}
\usage{
rgeopois(n, lambda, p)
}
\arguments{
\item{n}{size of sample.}

\item{lambda}{numeric parameter which must be numeric greater than  0.}

\item{p}{numeric parameter between 0 and 1.}
}
\value{
vector of samples generate from discrete geometric with Poisson hidden truncation distribution.
}
\description{
Random sample generating function for discrete Pareto distribution with parameters \eqn{\lambda > 0} and \eqn{p \in (0,1)}.
}
\details{
rgeopois generates random sample from discrete geometric with Poisson hidden truncation distribution.
}
\examples{
N <- rgeopois(20, lambda=2, p=0.3)
N

}
\references{
Amponsah, C. K. and  Kozubowski, T. J. (2020). Geometric with Poisson hidden truncation distribution.
\url{Yet to publish}
}
